+++ smartpark-busi/src/main/java/com/chris/smartpark/busi/entity/UserStationEntity.java
@@ -15,11 +15,11 @@ public class UserStationEntity  extends SysUpdateInfo  implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	//id
-	private Integer id;
+	private Long id;
 	//员工id
-	private Integer userId;
+	private Long userId;
 	//工位id
-	private Integer stationId;
+	private Long stationId;
 	//状态 0 废弃 1正常
 	private String status;
 	//员工工位IP
@@ -35,25 +35,25 @@ public class UserStationEntity  extends SysUpdateInfo  implements Serializable {
 	//扩展字段3
 	private String ext3;
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
-	public void setStationId(Integer stationId) {
+	public void setStationId(Long stationId) {
 		this.stationId = stationId;
 	}
 
-	public Integer getStationId() {
+	public Long getStationId() {
 		return stationId;
 	}
 	public void setStatus(String status) {
